modules/location_kafka_producer/requirements.txt
kafka-python
grpcio-tools 
grpcio